Refactor + split up DayTypeRepeat into small components + working on completing DayTypeRepeat
