+++ components/shared/repeat/day-type-repeat/DayTypeRepeat.js
@@ -3,15 +3,9 @@ import React, { Component } from 'react'
 import {
     View,
     Text,
-    TouchableHighlight,
     TouchableOpacity,
-    StyleSheet,
     TextInput,
-    DatePickerIOS,
     Keyboard,
-    Modal,
-    Picker,
-    TouchableWithoutFeedback,
     Dimensions,
     Animated,
     Easing,
@@ -21,10 +15,13 @@ import {
 
 import { styles } from './styles/styles'
 
+import RepeatValueHolder from './repeat-value-holder/RepeatValueHolder'
+import ChooseDayInWeekOption from './choose-day-in-week-option/ChooseDayInWeekOption'
+
+import RepeatEndOptionsHolder from './repeat-end-options-holder/RepeatEndOptionsHolder'
+
 import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome'
 import {
-    faRedoAlt,
-    faHourglassEnd,
     faFlag,
     faTimes,
     faCheck
@@ -33,8 +30,6 @@ import {
 const animation_duration = 250
 const easing = Easing.inOut(Easing.linear)
 const window_width = Dimensions.get("window").width
-const window_height = Dimensions.get("window").height
-const repeat_panel_height = 584
 const margin_bottom_of_last_row = 35
 
 export default class DayTypeRepeat extends React.PureComponent {
@@ -45,10 +40,6 @@ export default class DayTypeRepeat extends React.PureComponent {
     state = {
         selected_repeat_type: "days",
 
-        prev_repeat_type: "days",
-
-        is_picker_opened: false,
-
         repeat_input_value: "1",
 
         after_occurrence_value: "1",
@@ -58,67 +49,50 @@ export default class DayTypeRepeat extends React.PureComponent {
         should_animate_translate_y: true
     }
 
-    _onChangeAfterOccurrenceValue = (e) => {
-        if (e.nativeEvent.text.length === 0) {
+    _setRepeatType = (repeat_type) => {
+        this.setState({
+            selected_repeat_type: repeat_type
+        })
+    }
+
+    _onChangeRepeatInput = (e) => {
+        this.setState({
+            repeat_input_value: e.nativeEvent.text.replace(/[^0-9]/g, '')
+        })
+    }
+
+    _resetRepeatInput = () => {
+        if (this.state.repeat_input_value.length === 0) {
             this.setState({
-                after_occurrence_value: "1"
+                repeat_input_value: "1"
             })
         }
+    }
 
-        else {
+    _onChangeAfterOccurrenceValue = (e) => {
+        this.setState({
+            after_occurrence_value: e.nativeEvent.text.replace(/[^0-9]/g, '')
+        })
+    }
+
+    _resetAfterOccurrenceInput = () => {
+        if (this.state.after_occurrence_value.length === 0) {
             this.setState({
-                after_occurrence_value: e.nativeEvent.text.replace(/[^0-9]/g, '')
+                after_occurrence_value: "1"
             })
         }
     }
 
     _onChangeGoalValue = (e) => {
-        if (e.nativeEvent.text.length === 0) {
+        this.setState({
+            goal_value: e.nativeEvent.text.replace(/[^0-9]/g, '')
+        })
+    }
+
+    _resetGoalValueInput = () => {
+        if (this.state.goal_value.length === 0) {
             this.setState({
                 goal_value: "1"
-            })
-        }
-
-        else {
-            this.setState({
-                goal_value: e.nativeEvent.text.replace(/[^0-9]/g, '')
-            })
-        }
-    }
-
-    _changePickerValue = (itemValue, itemIndex) => {
-        this.setState(prevState => ({
-            selected_repeat_type: itemValue,
-            prev_repeat_type: prev_repeat_type.selected_repeat_type
-        }))
-    }
-
-    _openPicker = () => {
-        this.setState({
-            is_picker_opened: true
-        })
-    }
-
-    _closePicker = () => {
-        this.setState({
-            is_picker_opened: false
-        })
-    }
-
-    _chooseCancelPicker = () => {
-
-    }
-
-    _onRepeatInputChange = (e) => {
-        if (e.nativeEvent.text.length === 0) {
-            this.setState({
-                repeat_input_value: "1"
-            })
-        }
-
-        else {
-            this.setState({
-                repeat_input_value: e.nativeEvent.text.replace(/[^0-9]/g, '')
             })
         }
     }
@@ -146,13 +120,20 @@ export default class DayTypeRepeat extends React.PureComponent {
         ]).start()
     }
 
-    _keyboardWillShowHandler = (e) => {
+    _dontAnimateRepeatWhenFocusInput = () => {
+        this.setState({
+            should_animate_translate_y: false
+        })
     }
 
     _keyboardWillHideHandler = (e) => {
         this.setState({
             should_animate_translate_y: true
         })
+
+        this._resetRepeatInput()
+        this._resetAfterOccurrenceInput()
+        this._resetGoalValueInput()
     }
 
     close = () => {
@@ -163,31 +144,13 @@ export default class DayTypeRepeat extends React.PureComponent {
         this.props.hideAction()
     }
 
-    _chooseInput = () => {
-        if (this._text_input_ref) {
-            this._text_input_ref.focus()
-        }
-    }
-
-    _setRef = (r) => {
-        this._text_input_ref = r
-    }
-
-    _onFocus = () => {
-        this.setState({
-            should_animate_translate_y: false
-        })
-    }
-
     componentDidMount() {
         this.animateRepeat()
 
-        this._keyboardWillShowListener = Keyboard.addListener("keyboardWillShow", this._keyboardWillShowHandler)
         this._keyboardWillHideListener = Keyboard.addListener("keyboardWillHide", this._keyboardWillHideHandler)
     }
 
     componentWillUnmount() {
-        Keyboard.removeListener("keyboardWillShow", this._keyboardWillShowHandler)
         Keyboard.removeListener("keyboardWillHide", this._keyboardWillHideHandler)
     }
 
@@ -204,200 +167,24 @@ export default class DayTypeRepeat extends React.PureComponent {
                     overflow: "hidden"
                 }}
             >
-                <ScrollView
-                    keyboardDismissMode="on-drag"
-                    scrollEnabled={false}
+                <KeyboardAvoidingView
+                    behavior={"position"}
+                    enabled={this.state.should_animate_translate_y}
                 >
-
-                    <KeyboardAvoidingView
-                        behavior={"position"}
-                        enabled={this.state.should_animate_translate_y}
+                    <ScrollView
+                        keyboardDismissMode="on-drag"
+                        scrollEnabled={false}
                     >
-                        <View
-                            style={{
-                                flexDirection: "row",
-                                marginLeft: 30,
-                                marginTop: 35,
-                                alignItems: "center"
-                            }}
-                        >
-                            <FontAwesomeIcon
-                                icon={faRedoAlt}
-                                color="#2C2C2C"
-                                size={14}
-                            />
 
-                            <Text
-                                style={styles.title_text}
-                            >
-                                Repeat
-                            </Text>
-                        </View>
+                        <RepeatValueHolder
+                            repeat_input_value={this.state.repeat_input_value}
+                            _onChangeRepeatInput={this._onChangeRepeatInput}
+                            _dontAnimateRepeatWhenFocusInput={this._dontAnimateRepeatWhenFocusInput}
+                            selected_repeat_type={this.state.selected_repeat_type}
+                            _setRepeatType={this._setRepeatType}
+                        />
 
-                        <View
-                            style={{
-                                marginTop: 25,
-                                flexDirection: "row",
-                                alignItems: "center",
-                                marginLeft: 59,
-                            }}
-                        >
-                            <TouchableOpacity
-                                style={{
-                                    flexDirection: "row",
-                                }}
-
-                                onPress={this._chooseInput}
-                            >
-                                <Text
-                                    style={styles.every_option_text}
-                                >
-                                    Every
-                                </Text>
-
-                                <TextInput
-                                    style={styles.every_option_input}
-                                    maxLength={2}
-                                    keyboardType="numbers-and-punctuation"
-                                    value={this.state.repeat_input_value}
-                                    onChange={this._onRepeatInputChange}
-                                    ref={this._setRef}
-
-                                    onFocus={this._onFocus}
-                                />
-                            </TouchableOpacity>
-
-
-                            <TouchableOpacity
-                                style={styles.picker_button_container}
-
-                                onPress={this._openPicker}
-                            >
-                                <Text
-                                    style={styles.every_option_text}
-                                >
-                                    {this.state.selected_repeat_type}
-                                </Text>
-                            </TouchableOpacity>
-
-                            <Modal
-                                transparent={true}
-                                visible={this.state.is_picker_opened}
-                            >
-                                <View
-                                    style={{
-                                        flex: 1,
-                                        position: "relative"
-                                    }}
-                                >
-                                    <TouchableOpacity
-                                        style={{
-                                            flex: 1,
-                                            width: window_width,
-                                            backgroundColor: "#000000",
-                                            opacity: 0.2,
-                                        }}
-
-                                        onPress={this._closePicker}
-                                    >
-                                    </TouchableOpacity>
-
-                                    <View
-                                        style={{
-                                            position: "absolute",
-                                            height: 200,
-                                            bottom: 0,
-                                            left: 0,
-                                            right: 0,
-                                            backgroundColor: "white",
-                                            borderTopLeftRadius: 20,
-                                            borderTopRightRadius: 20,
-                                        }}
-                                    >
-                                        <View
-                                            style={{
-                                                flexDirection: "row",
-                                                justifyContent: "space-between",
-                                                alignItems: "center",
-                                                top: 20,
-                                                left: 0,
-                                                right: 0,
-                                                position: "absolute",
-                                                zIndex: 15,
-                                                height: 30,
-                                                paddingHorizontal: 30,
-                                            }}
-                                        >
-                                            <TouchableOpacity
-                                                style={{
-                                                    height: 30,
-                                                    justifyContent: "center",
-                                                    alignItems: "center"
-                                                }}
-
-                                                onPress={this._chooseCancelPicker}
-                                            >
-                                                <Text
-                                                    style={styles.picker_cancel_option_text}
-                                                >
-                                                    Cancel
-                                                </Text>
-                                            </TouchableOpacity>
-
-                                            <TouchableOpacity
-                                                style={{
-                                                    height: 30,
-                                                    justifyContent: "center",
-                                                    alignItems: "center"
-                                                }}
-                                            >
-                                                <Text
-                                                    style={styles.picker_done_option_text}
-                                                >
-                                                    Done
-                                                </Text>
-                                            </TouchableOpacity>
-                                        </View>
-
-                                        <Picker
-                                            selectedValue={this.state.selected_repeat_type}
-                                            onValueChange={this._changePickerValue}
-                                            itemStyle={styles.picker_value_text}
-                                            style={{
-                                                flex: 1,
-                                                justifyContent: "center",
-                                                zIndex: 10,
-                                                marginTop: 50,
-                                            }}
-                                        >
-                                            <Picker.Item
-                                                label="days"
-                                                value="days"
-                                            />
-                                            <Picker.Item
-                                                label="weeks"
-                                                value="weeks"
-                                            />
-                                            <Picker.Item
-                                                label="months"
-                                                value="months"
-                                            />
-                                        </Picker>
-                                    </View>
-                                </View>
-                            </Modal>
-                        </View>
-
-                        <View
-                            style={{
-                                marginTop: 25,
-                                flexDirection: "row",
-                                alignItems: "center",
-                                marginHorizontal: 30,
-                            }}
-                        >
-                            <ChooseDayInWeekOption />
-                        </View>
+                        <ChooseDayInWeekOption />
 
                         {/* Separating line */}
                         <View
@@ -419,7 +206,6 @@ export default class DayTypeRepeat extends React.PureComponent {
                         <GoalHolder
                             goal_value={this.state.goal_value}
                             _onChangeGoalValue={this._onChangeGoalValue}
-                            setPy={this.setPy}
                         />
 
                         <View
@@ -452,780 +238,11 @@ export default class DayTypeRepeat extends React.PureComponent {
                                 />
                             </TouchableOpacity>
                         </View>
-                    </KeyboardAvoidingView>
 
-                </ScrollView>
+                    </ScrollView>
+                </KeyboardAvoidingView>
+
             </Animated.View >
-        )
-    }
-}
-
-class ChooseDayInWeekOption extends React.PureComponent {
-
-    render() {
-        return (
-            <>
-                <DayInWeekOptionHolder
-                    index={0}
-                    day_in_week={"Mon"}
-                />
-                <DayInWeekOptionHolder
-                    index={1}
-                    day_in_week={"Tue"}
-                />
-                <DayInWeekOptionHolder
-                    index={2}
-                    day_in_week={"Wed"}
-                />
-                <DayInWeekOptionHolder
-                    index={3}
-                    day_in_week={"Thu"}
-                />
-                <DayInWeekOptionHolder
-                    index={4}
-                    day_in_week={"Fri"}
-                />
-                <DayInWeekOptionHolder
-                    index={5}
-                    day_in_week={"Sat"}
-                />
-                <DayInWeekOptionHolder
-                    index={6}
-                    day_in_week={"Sun"}
-                />
-            </>
-        )
-    }
-}
-
-
-class DayInWeekOptionHolder extends React.PureComponent {
-
-    handleIndex = (index) => {
-        if (index === 0) {
-            return styles.unchosen_left_end_day_in_week_container
-        }
-
-        else if (index === 6) {
-            return styles.unchosen_right_end_day_in_week_container
-        }
-        else {
-            return styles.unchosen_normal_day_in_week_container
-        }
-    }
-
-    state = {
-        container_style: this.handleIndex(this.props.index),
-        text_style: styles.unchosen_day_in_week_text,
-
-        is_chosen: false,
-    }
-
-    _onChooseDayInWeek = () => {
-        let { index } = this.props
-
-        this.setState(prevState => ({
-            is_chosen: !prevState.is_chosen
-        }), () => {
-            if (this.state.is_chosen) {
-                if (index === 0) {
-                    this.setState({
-                        container_style: styles.chosen_left_end_day_in_week_container,
-                        text_style: styles.chosen_day_in_week_text
-                    })
-                }
-
-                else if (index === 6) {
-                    this.setState({
-                        container_style: styles.chosen_right_end_day_in_week_container,
-                        text_style: styles.chosen_day_in_week_text
-                    })
-                }
-
-                else {
-                    this.setState({
-                        container_style: styles.chosen_normal_day_in_week_container,
-                        text_style: styles.chosen_day_in_week_text
-                    })
-                }
-            }
-
-            else {
-                if (index === 0) {
-                    this.setState({
-                        container_style: styles.unchosen_left_end_day_in_week_container,
-                        text_style: styles.unchosen_day_in_week_text
-                    })
-                }
-
-                else if (index === 6) {
-                    this.setState({
-                        container_style: styles.unchosen_right_end_day_in_week_container,
-                        text_style: styles.unchosen_day_in_week_text
-                    })
-                }
-
-                else {
-                    this.setState({
-                        container_style: styles.unchosen_normal_day_in_week_container,
-                        text_style: styles.unchosen_day_in_week_text
-                    })
-                }
-            }
-        })
-
-    }
-
-    render() {
-        return (
-            <>
-                {this.props.index === 0 ?
-                    <TouchableOpacity
-                        style={this.state.container_style}
-                        onPress={this._onChooseDayInWeek}
-                    >
-                        <Text
-                            style={this.state.text_style}
-                        >
-                            {this.props.day_in_week}
-                        </Text>
-                    </TouchableOpacity>
-
-                    :
-
-                    <>
-                        {this.props.index === 6 ?
-                            <TouchableOpacity
-                                style={this.state.container_style}
-                                onPress={this._onChooseDayInWeek}
-                            >
-                                <Text
-                                    style={this.state.text_style}
-                                >
-                                    {this.props.day_in_week}
-                                </Text>
-                            </TouchableOpacity>
-                            :
-
-                            <TouchableOpacity
-                                style={this.state.container_style}
-                                onPress={this._onChooseDayInWeek}
-                            >
-                                <Text
-                                    style={this.state.text_style}
-                                >
-                                    {this.props.day_in_week}
-                                </Text>
-                            </TouchableOpacity>
-                        }
-                    </>
-                }
-            </>
-        )
-    }
-}
-
-class RepeatEndOptionsHolder extends React.PureComponent {
-
-    state = {
-        current_index: 0,
-        last_index: -1
-    }
-
-    chooseEndOption = (index) => {
-        if (this.state.current_index !== index) {
-            this.setState(prevState => ({
-                current_index: index,
-                last_index: prevState.current_index
-            }))
-        }
-    }
-
-    render() {
-        return (
-            <View>
-                <View
-                    style={{
-                        flexDirection: "row",
-                        marginLeft: 30,
-                        marginTop: 25,
-                        alignItems: "center"
-                    }}
-                >
-                    <FontAwesomeIcon
-                        icon={faHourglassEnd}
-                        color="#2C2C2C"
-                        size={14}
-                    />
-
-                    <Text
-                        style={styles.title_text}
-                    >
-                        Repeat end
-                    </Text>
-                </View>
-
-                <RepeatEndNeverOptionRow
-                    index={0}
-                    chooseEndOption={this.chooseEndOption}
-                    current_index={this.state.current_index}
-                    last_index={this.state.last_index}
-                />
-
-                <RepeatEndOnOptionRow
-                    index={1}
-                    chooseEndOption={this.chooseEndOption}
-                    current_index={this.state.current_index}
-                    last_index={this.state.last_index}
-                />
-
-                <RepeatEndAfterOptionRow
-                    index={2}
-                    chooseEndOption={this.chooseEndOption}
-                    current_index={this.state.current_index}
-                    last_index={this.state.last_index}
-                    {...this.props}
-                />
-            </View>
-        )
-    }
-}
-
-class RepeatEndNeverOptionRow extends React.Component {
-
-    shouldComponentUpdate(nextProps, nextState) {
-        return (this.props.index === nextProps.current_index && this.props.current_index !== nextProps.current_index)
-            || (this.props.index === nextProps.last_index && this.props.last_index !== nextProps.last_index)
-    }
-
-    _chooseEndOption = () => {
-        this.props.chooseEndOption(this.props.index)
-    }
-
-    render() {
-        let text_style = styles.unchosen_every_option_text,
-            button_container = styles.repeat_end_chosen_button_container_deactivated,
-            activated = this.props.index === this.props.current_index
-
-        if (this.props.index === this.props.current_index) {
-            text_style = styles.every_option_text
-            button_container = styles.repeat_end_chosen_button_container
-        }
-
-        else if (this.props.index === this.props.last_index) {
-            text_style = styles.unchosen_every_option_text
-            button_container = styles.repeat_end_chosen_button_container_deactivated
-        }
-
-        return (
-            <TouchableOpacity
-                style={{
-                    marginTop: 25,
-                    marginLeft: 59,
-                    marginRight: 30,
-                }}
-
-                onPress={this._chooseEndOption}
-            >
-                <View
-                    style={{
-                        flexDirection: "row",
-                        alignItems: "center",
-                        justifyContent: "space-between"
-                    }}
-                >
-                    <Text
-                        style={text_style}
-                    >
-                        Never
-                        </Text>
-
-                    <View
-                        style={button_container}
-                    >
-                        {activated ?
-                            <View
-                                style={styles.repeat_end_chosen_button_activated}
-                            >
-
-                            </View>
-                            :
-                            null
-                        }
-                    </View>
-                </View>
-            </TouchableOpacity>
-        )
-    }
-}
-
-class RepeatEndOnOptionRow extends React.Component {
-
-    month_names = ["January", "Febuary", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"]
-
-    date = new Date()
-
-    state = {
-        is_date_picker_chosen: false,
-        day: this.date.getDate().toString(),
-        month: this.date.getMonth().toString(),
-        year: this.date.getFullYear().toString()
-    }
-
-    shouldComponentUpdate(nextProps, nextState) {
-        return (this.props.index === nextProps.current_index && this.props.current_index !== nextProps.current_index)
-            || (this.props.index === nextProps.last_index && this.props.last_index !== nextProps.last_index)
-            || (this.state !== nextState)
-    }
-
-    _chooseEndOption = () => {
-        this.props.chooseEndOption(this.props.index)
-        this.setState({
-            is_date_picker_chosen: true
-        })
-    }
-
-    closeDatePicker = () => {
-        this.setState({
-            is_date_picker_chosen: false
-        })
-    }
-
-    _changeDayPickerValue = (itemValue, itemIndex) => {
-        let year = parseInt(this.state.year),
-            month = parseInt(this.state.month),
-            day = parseInt(itemValue)
-
-        let date = new Date(year, month, day)
-
-        if (date.getMonth() !== month) {
-            date = new Date(year, month + 1, 0)
-        }
-
-        this.setState({
-            day: date.getDate().toString()
-        })
-    }
-
-    _changeMonthPickerValue = (itemValue, itemIndex) => {
-        let year = parseInt(this.state.year),
-            month = parseInt(itemValue),
-            day = parseInt(this.state.day)
-
-        let date = new Date(year, month, day)
-
-        if (date.getMonth() !== month) {
-            date = new Date(year, month + 1, 0)
-        }
-
-        this.setState({
-            month: itemValue,
-            day: date.getDate().toString()
-        })
-    }
-
-    _changeYearPickerValue = (itemValue, itemIndex) => {
-        let year = parseInt(this.state.year),
-            month = parseInt(itemValue),
-            day = parseInt(this.state.day)
-
-        let date = new Date(year, month, day)
-
-        if (date.getMonth() !== month) {
-            date = new Date(year, month + 1, 0)
-        }
-
-        this.setState({
-            year: itemValue,
-            day: date.getDate().toString()
-
-        })
-    }
-
-
-    render() {
-        let text_style = styles.unchosen_every_option_text,
-            button_container = styles.repeat_end_chosen_button_container_deactivated,
-            activated = this.props.index === this.props.current_index
-
-        if (this.props.index === this.props.current_index) {
-            text_style = styles.every_option_text
-            button_container = styles.repeat_end_chosen_button_container
-        }
-
-        else if (this.props.index === this.props.last_index) {
-            text_style = styles.unchosen_every_option_text
-            button_container = styles.repeat_end_chosen_button_container_deactivated
-        }
-
-        return (
-            <>
-                <TouchableOpacity
-                    style={{
-                        marginTop: 25,
-                        marginLeft: 59,
-                        marginRight: 30,
-                    }}
-
-                    onPress={this._chooseEndOption}
-                >
-                    <View
-                        style={{
-                            flexDirection: "row",
-                            alignItems: "center",
-                            justifyContent: "space-between"
-                        }}
-                    >
-                        <View
-                            style={{
-                                flexDirection: "row",
-                                alignItems: "center",
-                            }}
-                        >
-                            <Text
-                                style={text_style}
-                            >
-                                On
-                        </Text>
-
-                            <View
-                                style={{
-                                    marginLeft: 20,
-                                }}
-                            >
-                                <Text
-                                    style={text_style}
-                                >
-                                    {`${this.state.day} ${this.month_names[this.state.month]} ${this.state.year}`}
-                                </Text>
-                            </View>
-                        </View>
-
-                        <View
-                            style={button_container}
-                        >
-                            {activated ?
-                                <View
-                                    style={styles.repeat_end_chosen_button_activated}
-                                >
-
-                                </View>
-                                :
-                                null
-                            }
-                        </View>
-                    </View>
-                </TouchableOpacity>
-
-                <Modal
-                    transparent={true}
-                    visible={this.state.is_date_picker_chosen}
-                >
-                    <View
-                        style={{
-                            flex: 1,
-                            position: "relative"
-                        }}
-                    >
-                        <TouchableOpacity
-                            style={{
-                                flex: 1,
-                                width: window_width,
-                                backgroundColor: "black",
-                                opacity: 0.2
-                            }}
-
-                            onPress={this.closeDatePicker}
-                        >
-
-                        </TouchableOpacity>
-
-                        <DatePickerWheel
-                            day={this.state.day}
-                            month={this.state.month}
-                            year={this.state.year}
-                            _changeDayPickerValue={this._changeDayPickerValue}
-                            _changeMonthPickerValue={this._changeMonthPickerValue}
-                            _changeYearPickerValue={this._changeYearPickerValue}
-                        />
-                    </View>
-                </Modal>
-            </>
-        )
-    }
-}
-
-class DatePickerWheel extends React.PureComponent {
-
-    render() {
-        return (
-            <View
-                style={{
-                    position: "absolute",
-                    height: 200,
-                    bottom: 0,
-                    left: 0,
-                    right: 0,
-                    backgroundColor: "white",
-                    justifyContent: "center",
-                    borderTopLeftRadius: 20,
-                    borderTopRightRadius: 20,
-                }}
-            >
-                <View
-                    style={{
-                        flex: 1,
-                        flexDirection: "row",
-                        alignItems: "center",
-                        justifyContent: "center"
-                    }}
-                >
-                    <DayPickerValueHolder
-                        day={this.props.day}
-                        _changeDayPickerValue={this.props._changeDayPickerValue}
-                    />
-                    <MonthPickerValueHolder
-                        month={this.props.month}
-                        _changeMonthPickerValue={this.props._changeMonthPickerValue}
-                    />
-                    <YearPickerValueHolder
-                        year={this.props.year}
-                        _changeYearPickerValue={this.props._changeYearPickerValue}
-                    />
-                </View>
-            </View>
-        )
-    }
-}
-
-class DayPickerValueHolder extends React.PureComponent {
-
-    render() {
-        return (
-            <Picker
-                selectedValue={this.props.day}
-                onValueChange={this.props._changeDayPickerValue}
-                itemStyle={styles.picker_value_text}
-                style={{
-                    height: 50,
-                    flex: 1,
-                    justifyContent: "center"
-                }}
-            >
-                <Picker.Item label="1" value="1" />
-                <Picker.Item label="2" value="2" />
-                <Picker.Item label="3" value="3" />
-                <Picker.Item label="4" value="4" />
-                <Picker.Item label="5" value="5" />
-                <Picker.Item label="6" value="6" />
-                <Picker.Item label="7" value="7" />
-                <Picker.Item label="8" value="8" />
-                <Picker.Item label="9" value="9" />
-                <Picker.Item label="10" value="10" />
-                <Picker.Item label="11" value="11" />
-                <Picker.Item label="12" value="12" />
-                <Picker.Item label="13" value="13" />
-                <Picker.Item label="14" value="14" />
-                <Picker.Item label="15" value="15" />
-                <Picker.Item label="16" value="16" />
-                <Picker.Item label="17" value="17" />
-                <Picker.Item label="18" value="18" />
-                <Picker.Item label="19" value="19" />
-                <Picker.Item label="20" value="20" />
-                <Picker.Item label="21" value="21" />
-                <Picker.Item label="22" value="22" />
-                <Picker.Item label="23" value="23" />
-                <Picker.Item label="24" value="24" />
-                <Picker.Item label="25" value="25" />
-                <Picker.Item label="26" value="26" />
-                <Picker.Item label="27" value="27" />
-                <Picker.Item label="28" value="28" />
-                <Picker.Item label="29" value="29" />
-                <Picker.Item label="30" value="30" />
-                <Picker.Item label="31" value="31" />
-            </Picker>
-        )
-    }
-}
-
-class MonthPickerValueHolder extends React.PureComponent {
-
-    render() {
-        return (
-            <Picker
-                selectedValue={this.props.month}
-                onValueChange={this.props._changeMonthPickerValue}
-                itemStyle={styles.picker_value_text}
-                style={{
-                    height: 50,
-                    flex: 1,
-                    justifyContent: "center"
-                }}
-            >
-                <Picker.Item label="January" value="0" />
-                <Picker.Item label="Febuary" value="1" />
-                <Picker.Item label="March" value="2" />
-                <Picker.Item label="April" value="3" />
-                <Picker.Item label="May" value="4" />
-                <Picker.Item label="June" value="5" />
-                <Picker.Item label="July" value="6" />
-                <Picker.Item label="August" value="7" />
-                <Picker.Item label="September" value="8" />
-                <Picker.Item label="October" value="9" />
-                <Picker.Item label="November" value="10" />
-                <Picker.Item label="December" value="11" />
-            </Picker>
-        )
-    }
-}
-
-class YearPickerValueHolder extends React.PureComponent {
-    year = new Date().getFullYear()
-    render() {
-        return (
-            <Picker
-                selectedValue={this.props.year}
-                onValueChange={this.props._changeYearPickerValue}
-                itemStyle={styles.picker_value_text}
-                style={{
-                    height: 50,
-                    flex: 1,
-                    justifyContent: "center"
-                }}
-
-            >
-                <Picker.Item label={`${this.year}`} value={`${this.year}`} />
-                <Picker.Item label={`${this.year + 1}`} value={`${this.year + 1}`} />
-                <Picker.Item label={`${this.year + 2}`} value={`${this.year + 2}`} />
-                <Picker.Item label={`${this.year + 3}`} value={`${this.year + 3}`} />
-                <Picker.Item label={`${this.year + 4}`} value={`${this.year + 4}`} />
-            </Picker>
-        )
-    }
-}
-
-
-class RepeatEndAfterOptionRow extends React.Component {
-
-    state = {
-        is_text_input_readable: false
-    }
-
-    shouldComponentUpdate(nextProps, nextState) {
-        return (this.props.index === nextProps.current_index && this.props.current_index !== nextProps.current_index)
-            || (this.props.index === nextProps.last_index && this.props.last_index !== nextProps.last_index)
-            || (this.props.after_occurrence_value !== nextProps.after_occurrence_value)
-    }
-
-    _chooseEndOption = () => {
-        this.props.chooseEndOption(this.props.index)
-        this.setState({
-            is_text_input_readable: true
-        }, () => {
-            this._text_input_ref.focus()
-        })
-    }
-
-    _setTextInputRef = (r) => {
-        this._text_input_ref = r
-    }
-
-    componentDidUpdate(prevProps, prevState) {
-        if (this.props.index === this.props.last_index && this.props.last_index !== prevProps.last_index) {
-            this.setState({
-                is_text_input_readable: false,
-            }, () => {
-                this._text_input_ref.blur()
-            })
-        }
-    }
-
-    render() {
-        let text_style = styles.unchosen_every_option_text,
-            button_container = styles.repeat_end_chosen_button_container_deactivated,
-            activated = this.props.index === this.props.current_index,
-            input_text_style = styles.unchosen_every_option_input
-
-        if (this.props.index === this.props.current_index) {
-            text_style = styles.every_option_text
-            button_container = styles.repeat_end_chosen_button_container
-            input_text_style = styles.every_option_input
-        }
-
-        else if (this.props.index === this.props.last_index) {
-            text_style = styles.unchosen_every_option_text
-            button_container = styles.repeat_end_chosen_button_container_deactivated
-            input_text_style = styles.unchosen_every_option_input
-        }
-
-        return (
-            <TouchableOpacity
-                style={{
-                    marginTop: 25,
-                    marginLeft: 59,
-                    marginRight: 30,
-                }}
-
-                onPress={this._chooseEndOption}
-            >
-                <View
-                    style={{
-                        flexDirection: "row",
-                        alignItems: "center",
-                        justifyContent: "space-between"
-                    }}
-                >
-                    <View
-                        style={{
-                            flexDirection: "row",
-                            alignItems: "center",
-                        }}
-                    >
-                        <Text
-                            style={text_style}
-                        >
-                            After
-                        </Text>
-
-                        <TextInput
-                            style={input_text_style}
-                            editable={this.state.is_text_input_readable}
-                            maxLength={2}
-                            keyboardType="numbers-and-punctuation"
-                            placeholder="1"
-                            value={this.props.after_occurrence_value}
-                            onChange={this.props._onChangeAfterOccurrenceValue}
-                            ref={this._setTextInputRef}
-                        />
-
-                        <View
-                            style={{
-                                marginLeft: 20,
-                            }}
-                        >
-                            <Text
-                                style={text_style}
-                            >
-                                occurrences
-                            </Text>
-                        </View>
-                    </View>
-
-                    <View
-                        style={button_container}
-                    >
-                        {activated ?
-                            <View
-                                style={styles.repeat_end_chosen_button_activated}
-                            >
-
-                            </View>
-                            :
-                            null
-                        }
-                    </View>
-                </View>
-            </TouchableOpacity>
         )
     }
 }
@@ -1285,6 +302,7 @@ class GoalHolder extends React.PureComponent {
                         value={this.props.goal_value}
                         onChange={this.props._onChangeGoalValue}
                         ref={this._setRef}
+                        autoCorrect={false}
                     />
 
                     <View
