+++ components/shared/repeat/day-type-repeat/repeat-end-options-holder/styles/styles.js
@@ -1,0 +1,17 @@
+import {
+    StyleSheet,
+
+} from 'react-native'
+
+import * as CommonStyle from '../../../../../shared/styles/style'
+
+export const styles = StyleSheet.create({
+    title_text: {
+        fontFamily: CommonStyle.sf_ui_display_light_font,
+        fontSize: 18,
+        lineHeight: 21,
+        letterSpacing: -0.02,
+        color: CommonStyle.text_icon_colors.ti_1,
+        marginLeft: 15,
+    },
+})
