+++ components/shared/repeat/day-type-repeat/repeat-value-holder/RepeatValueHolder.Container.js
@@ -1,0 +1,12 @@
+import {connect} from 'react-redux'
+
+import RepeatValueHolder from './RepeatValueHolder'
+
+const mapStateToProps = (state) => ({
+    currentDayTask: state.get("currentDayTask"),
+})
+
+export default connect(
+    mapStateToProps,
+    mapDispatchToProps
+)(RepeatValueHolder)
